# frontend/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "BookingCreate": {
        "properties": {
          "num_seats": {
            "default": 1,
            "minimum": 1.0,
            "title": "Num Seats",
            "type": "integer"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "ride_id": {
            "format": "uuid",
            "title": "Ride Id",
            "type": "string"
          }
        },
        "required": [
          "ride_id",
          "request_id"
        ],
        "title": "BookingCreate",
        "type": "object"
      },
...
      "BookingResponse": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "num_seats": {
            "title": "Num Seats",
            "type": "integer"
          },
          "passenger_id": {
            "format": "uuid",
            "title": "Passenger Id",
            "type": "string"
          },
          "passenger_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passenger Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "ride_id": {
            "format": "uuid",
            "title": "Ride Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          }
        },
        "required": [
          "booking_id",
          "ride_id",
          "passenger_id",
          "num_seats",
          "status",
          "created_at"
        ],
        "title": "BookingResponse",
        "type": "object"
      },
...
      "DriverActiveResponse": {
        "properties": {
          "rides": {
            "items": {
              "$ref": "#/components/schemas/RideWithPassengersItem"
            },
            "title": "Rides",
            "type": "array"
          }
        },
        "required": [
          "rides"
        ],
        "title": "DriverActiveResponse",
        "type": "object"
      },
      "DriverHistoryResponse": {
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "rides": {
            "items": {
              "$ref": "#/components/schemas/RideWithPassengersItem"
            },
            "title": "Rides",
            "type": "array"
          }
        },
        "required": [
          "rides"
        ],
        "title": "DriverHistoryResponse",
        "type": "object"
      },
...
      "DriverLocationReport": {
        "description": "Request body for driver location report (POST /bookings/{booking_id}/location).",
        "properties": {
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "DriverLocationReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PassengerActiveResponse": {
        "properties": {
          "bookings": {
            "items": {
              "$ref": "#/components/schemas/PassengerBookingSummaryItem"
            },
            "title": "Bookings",
            "type": "array"
          }
        },
        "required": [
          "bookings"
        ],
        "title": "PassengerActiveResponse",
        "type": "object"
      },
...
      "PassengerHistoryResponse": {
        "properties": {
          "bookings": {
            "items": {
              "$ref": "#/components/schemas/PassengerBookingSummaryItem"
            },
            "title": "Bookings",
            "type": "array"
          },
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "bookings"
        ],
        "title": "PassengerHistoryResponse",
        "type": "object"
      },
...
      "RideManifestResponse": {
        "properties": {
          "available_seats_left": {
            "title": "Available Seats Left",
            "type": "integer"
          },
          "confirmed_total": {
            "description": "כל שורות confirmed לנסיעה (COUNT)",
            "minimum": 0.0,
            "title": "Confirmed Total",
            "type": "integer"
          },
          "manifest_truncated": {
            "description": "True אם confirmed_total + pending_total חורגים מ-manifest row limit והרשימה נחתכה",
            "title": "Manifest Truncated",
            "type": "boolean"
          },
          "passengers": {
            "items": {
              "$ref": "#/components/schemas/BookingManifestItem"
            },
            "title": "Passengers",
            "type": "array"
          },
          "pending_total": {
            "description": "כל שורות pending לנסיעה (COUNT)",
            "minimum": 0.0,
            "title": "Pending Total",
            "type": "integer"
          },
          "ride_id": {
            "format": "uuid",
            "title": "Ride Id",
            "type": "string"
          },
          "total_confirmed_passengers": {
            "description": "תואם confirmed_total לתאימות לאחור; ספירת מאושרים ב-DB לאורך הנסיעה",
            "minimum": 0.0,
            "title": "Total Confirmed Passengers",
            "type": "integer"
          }
        },
        "required": [
          "ride_id",
          "confirmed_total",
          "pending_total",
          "manifest_truncated",
          "total_confirmed_passengers",
          "available_seats_left",
          "passengers"
        ],
        "title": "RideManifestResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/bookings/{booking_id}": {
      "get": {
        "operationId": "admin_booking_by_id_api_v1_admin_bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Booking By Id",
        "tags": [
          "Admin",
          "Admin"
        ]
      }
    },
...
    "/api/v1/bookings/driver-summary/active": {
      "get": {
        "description": "Active driver rides only (paginated counterpart uses /driver-summary/history).",
        "operationId": "get_driver_active_summary_api_v1_bookings_driver_summary_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverActiveResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver Active Summary",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/driver-summary/history": {
      "get": {
        "operationId": "get_driver_history_summary_api_v1_bookings_driver_summary_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver History Summary",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/my-bookings": {
      "get": {
        "operationId": "get_user_bookings_api_v1_bookings_my_bookings_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingResponse"
                  },
                  "title": "Response Get User Bookings Api V1 Bookings My Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Bookings",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/passenger-summary/active": {
      "get": {
        "operationId": "get_passenger_active_summary_api_v1_bookings_passenger_summary_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerActiveResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Passenger Active Summary",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/passenger-summary/history": {
      "get": {
        "operationId": "get_passenger_history_summary_api_v1_bookings_passenger_summary_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Passenger History Summary",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/request-to-join": {
      "post": {
        "description": "Request to join a ride. request_id must belong to the authenticated user.",
        "operationId": "request_to_join_api_v1_bookings_request_to_join_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request To Join",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/ride/{ride_id}/manifest": {
      "get": {
        "operationId": "get_ride_manifest_api_v1_bookings_ride__ride_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ride Manifest",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/ride/{ride_id}/pending": {
      "get": {
        "operationId": "get_pending_requests_api_v1_bookings_ride__ride_id__pending_get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingResponse"
                  },
                  "title": "Response Get Pending Requests Api V1 Bookings Ride  Ride Id  Pending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Pending Requests",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}": {
      "get": {
        "operationId": "get_booking_api_v1_bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Booking",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/approve": {
      "patch": {
        "operationId": "approve_booking_api_v1_bookings__booking_id__approve_patch",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Booking",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "cancel_booking_api_v1_bookings__booking_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Booking",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/location": {
      "post": {
        "description": "Driver reports location during the ride. Broadcast to confirmed passengers over WebSocket.\nRequires: ride driver, ride in active status.",
        "operationId": "report_driver_location_api_v1_bookings__booking_id__location_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverLocationReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Report Driver Location",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/passenger-location": {
      "post": {
        "description": "Passenger reports location during the ride. Broadcast to driver on ride_{ride_id}:passenger_locations.\nRequires: the booking’s passenger.",
        "operationId": "report_passenger_location_api_v1_bookings__booking_id__passenger_location_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassengerLocationReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Report Passenger Location",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/reject": {
      "patch": {
        "operationId": "reject_booking_api_v1_bookings__booking_id__reject_patch",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Booking",
        "tags": [
          "Bookings",
          "Bookings"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://127.0.0.1:8000"
    }
  ]
}